+++ svg/js/src/main/scala/doodle/svg/algebra/Algebra.scala
@@ -3,7 +3,10 @@ package svg
 package algebra
 
 import cats._
+import doodle.core.BoundingBox
+import doodle.core.font.Font
 import doodle.language.Basic
+import doodle.svg.effect.Canvas
 
 trait JsAlgebraModule
     extends AlgebraModule
@@ -11,15 +14,22 @@ trait JsAlgebraModule
     with PathModule
     with ShapeModule
     with SvgModule
+    with TextModule
     with JsBase {
   type Algebra[F[_]] = doodle.algebra.Algebra[F]
     with doodle.interact.algebra.MouseOver[F]
+    with doodle.algebra.Text[F]
     with Basic[F]
 
   final class JsAlgebra(
+      val canvas: Canvas,
       val applyF: Apply[SvgResult],
       val functorF: Functor[SvgResult]
   ) extends BaseAlgebra
+      with HasTextBoundingBox
       with MouseOver
-  val algebraInstance = new JsAlgebra(Svg.svgResultApply, Svg.svgResultApply)
+      with Text {
+    def textBoundingBox(text: String, font: Font): BoundingBox =
+      canvas.textBoundingBox(text, font)
+  }
 }