+++ svg/js/src/main/scala/doodle/svg/effect/Canvas.scala
@@ -3,18 +3,24 @@ package svg
 package effect
 
 import cats.effect.IO
-import doodle.core.{BoundingBox, Color,Point,Transform}
+import doodle.core.{BoundingBox, Color, Point, Transform}
+import doodle.core.font.Font
 import monix.reactive.Observable
 import monix.reactive.subjects.PublishSubject
 import org.scalajs.dom
 import scalatags.JsDom
+import scalatags.JsDom.svgTags
+import scalatags.JsDom.svgAttrs
 
-final case class Canvas(target: dom.Node,
-                        frame: Frame,
-                        background: Option[Color]) {
+final case class Canvas(
+    target: dom.Node,
+    frame: Frame,
+    background: Option[Color]
+) {
   import JsDom.all.{Tag => _, _}
 
-  val algebra: Algebra[Drawing] = algebraInstance
+  val algebra: Algebra[Drawing] =
+    new js.JsAlgebra(this, Svg.svgResultApply, Svg.svgResultApply)
 
   val redraw: Observable[Int] = {
     val subject = PublishSubject[Int]()
@@ -54,6 +60,7 @@ final case class Canvas(target: dom.Node,
 
   private var currentBB: BoundingBox = _
   private var svgRoot: dom.Node = _
+
   /** Get the root <svg> node, creating one if needed. */
   def svgRoot(bb: BoundingBox): dom.Node = {
     def addCallback(tag: Tag, tx: Transform): Tag =
@@ -88,6 +95,16 @@ final case class Canvas(target: dom.Node,
     }
   }
 
+  def textBoundingBox(text: String, font: Font): BoundingBox = {
+    // Create an invisible SVG element to measure the text size and delete it
+    // after use
+    val elt = target.appendChild(svgTags.svg(svgAttrs.display:="none").render)
+    val txt = elt.appendChild(Svg.textTag(text, font).render)
+    val bb = txt.asInstanceOf[dom.svg.Text].getBBox()
+
+    BoundingBox.centered(bb.width, bb.height)
+  }
+
   def render[A](picture: Picture[A]): IO[A] = {
     for {
       result <- Svg.renderWithoutRootTag[Algebra, A](algebra, picture)
@@ -102,9 +119,10 @@ final case class Canvas(target: dom.Node,
 object Canvas {
   def fromFrame(frame: Frame): Canvas = {
     val target = dom.document.getElementById(frame.id)
-    if(target == null) {
+    if (target == null) {
       throw new java.util.NoSuchElementException(
-        s"Doodle SVG Canvas could not be created, as could not find a DOM element with the requested id ${frame.id}")
+        s"Doodle SVG Canvas could not be created, as could not find a DOM element with the requested id ${frame.id}"
+      )
     } else Canvas(target, frame, frame.background)
   }
 }